% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irf.R
\name{autoplot.irf_long}
\alias{autoplot.irf_long}
\title{Plotting irf from long format}
\usage{
\method{autoplot}{irf_long}(object, filter_shock = NULL,
  filter_var = NULL, ...)
}
\arguments{
\item{object}{of class irf_long}

\item{filter_shock}{choose shocks to be plotted}

\item{filter_var}{choose variables to be plotted}

\item{...}{pass to facet_wrap}
}
\description{
Plotting irf from long format
}
