% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr_dynare.R
\name{clean_dynare}
\alias{clean_dynare}
\title{Clean the imported matlab .mat matrix}
\usage{
clean_dynare(x)
}
\arguments{
\item{x}{an imported *_results.mat file}
}
\description{
Clean the imported matlab .mat matrix
}
