% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irf.R
\name{irf_long}
\alias{irf_long}
\title{Extract the irfs into long format}
\usage{
irf_long(x, regexpr = "\\\\.e\\\\.")
}
\arguments{
\item{x}{a result object after cleanning.}

\item{regexpr}{regular expression for seperation}
}
\description{
Extract the irfs into long format
}
