% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-mat.R
\name{mat2r}
\alias{mat2r}
\title{Imports from matlab to R}
\usage{
mat2r(matfile = NULL, path = ".")
}
\arguments{
\item{matfile}{a .mat file.}

\item{path}{the specified path to find the file.}
}
\description{
Imports from matlab to R
}
